# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: F1 Live Telecronaca API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /drivers:
    get:
      operationId: listDrivers
      tags: [drivers]
      summary: List all drivers
      responses:
        "200":
          description: List of drivers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Driver"
    post:
      operationId: createDriver
      tags: [drivers]
      summary: Create a driver (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DriverInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
  /drivers/{id}:
    get:
      operationId: getDriver
      tags: [drivers]
      summary: Get driver by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Driver
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
    patch:
      operationId: updateDriver
      tags: [drivers]
      summary: Update a driver (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DriverUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
    delete:
      operationId: deleteDriver
      tags: [drivers]
      summary: Delete a driver (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /races/{id}/laps:
    get:
      operationId: listLaps
      tags: [laps]
      summary: List lap times for a race
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lap times
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LapTime"
    post:
      operationId: addLap
      tags: [laps]
      summary: Add a lap time (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LapTimeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LapTime"
  /races/{id}/laps/latest:
    get:
      operationId: getLatestPositions
      tags: [laps]
      summary: Get latest timing tower positions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Current positions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DriverPosition"
  /races/{id}/commentary:
    get:
      operationId: listCommentary
      tags: [commentary]
      summary: List commentary for a race
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Commentary entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Commentary"
    post:
      operationId: addCommentary
      tags: [commentary]
      summary: Add commentary entry (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentaryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Commentary"
  /commentary/{id}:
    delete:
      operationId: deleteCommentary
      tags: [commentary]
      summary: Delete a commentary entry (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /races/{id}/race-control:
    get:
      operationId: listRaceControl
      tags: [race-control]
      summary: List race control messages for a race
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Race control messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RaceControlMessage"
    post:
      operationId: addRaceControl
      tags: [race-control]
      summary: Post race control message (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RaceControlInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RaceControlMessage"
...
components:
  schemas:
...
    Driver:
      type: object
      required: [id, name, abbreviation, number, nationality, teamId]
      properties:
        id:
          type: integer
        name:
          type: string
        abbreviation:
          type: string
        number:
          type: integer
        nationality:
          type: string
        teamId:
          type: integer
        teamName:
          type: ["string", "null"]
        teamColor:
          type: ["string", "null"]
    DriverInput:
      type: object
      required: [name, abbreviation, number, nationality, teamId]
      properties:
        name:
          type: string
        abbreviation:
          type: string
        number:
          type: integer
        nationality:
          type: string
        teamId:
          type: integer
    DriverUpdate:
      type: object
      properties:
        name:
          type: string
        abbreviation:
          type: string
        number:
          type: integer
        nationality:
          type: string
        teamId:
          type: integer
...
    LapTime:
      type: object
      required: [id, raceId, driverId, lap, position, lapTimeMs]
      properties:
        id:
          type: integer
        raceId:
          type: integer
        driverId:
          type: integer
        lap:
          type: integer
        position:
          type: integer
        lapTimeMs:
          type: integer
        sector1Ms:
          type: ["integer", "null"]
        sector2Ms:
          type: ["integer", "null"]
        sector3Ms:
          type: ["integer", "null"]
        pitStop:
          type: boolean
        pitStopDurationMs:
          type: ["integer", "null"]
        isPurpleSector1:
          type: boolean
        isPurpleSector2:
          type: boolean
        isPurpleSector3:
          type: boolean
        isFastestLap:
          type: boolean
    LapTimeInput:
      type: object
      required: [driverId, lap, position, lapTimeMs]
      properties:
        driverId:
          type: integer
        lap:
          type: integer
        position:
          type: integer
        lapTimeMs:
          type: integer
        sector1Ms:
          type: ["integer", "null"]
        sector2Ms:
          type: ["integer", "null"]
        sector3Ms:
          type: ["integer", "null"]
        pitStop:
          type: boolean
        pitStopDurationMs:
          type: ["integer", "null"]
    DriverPosition:
      type: object
      required: [position, driverId, driverName, driverAbbr, teamName, teamColor, lap, gapToLeaderMs, intervalMs]
      properties:
        position:
          type: integer
        driverId:
          type: integer
        driverName:
          type: string
        driverAbbr:
          type: string
        driverNumber:
          type: integer
        teamName:
          type: string
        teamColor:
          type: string
        lap:
          type: integer
        gapToLeaderMs:
          type: integer
        intervalMs:
          type: integer
        lastLapTimeMs:
          type: ["integer", "null"]
        currentTire:
          type: ["string", "null"]
        tireLaps:
          type: ["integer", "null"]
        pitStops:
          type: integer
        dnf:
          type: boolean
        dnfReason:
          type: ["string", "null"]
    Commentary:
      type: object
      required: [id, raceId, message, type, timestamp]
      properties:
        id:
          type: integer
        raceId:
          type: integer
        message:
          type: string
        type:
          type: string
          enum: [normal, safety_car, vsc, flag, radio, incident, fastest_lap, overtake, pit_stop, dnf]
        lap:
          type: ["integer", "null"]
        driverName:
          type: ["string", "null"]
        timestamp:
          type: string
    CommentaryInput:
      type: object
      required: [message, type]
      properties:
        message:
          type: string
        type:
          type: string
          enum: [normal, safety_car, vsc, flag, radio, incident, fastest_lap, overtake, pit_stop, dnf]
        lap:
          type: ["integer", "null"]
        driverName:
          type: ["string", "null"]
    RaceControlMessage:
      type: object
      required: [id, raceId, message, category, timestamp]
      properties:
        id:
          type: integer
        raceId:
          type: integer
        message:
          type: string
        category:
          type: string
          enum: [flag, safety_car, vsc, drs, track, other]
        flag:
          type: ["string", "null"]
          enum: ["green", "yellow", "red", "chequered", "white", "black_orange", null]
        lap:
          type: ["integer", "null"]
        timestamp:
          type: string
    RaceControlInput:
      type: object
      required: [message, category]
      properties:
        message:
          type: string
        category:
          type: string
          enum: [flag, safety_car, vsc, drs, track, other]
        flag:
          type: ["string", "null"]
          enum: ["green", "yellow", "red", "chequered", "white", "black_orange", null]
        lap:
          type: ["integer", "null"]
...