# lib/api-zod/src/generated/types/commentaryInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * F1 Live Telecronaca API
 * OpenAPI spec version: 0.1.0
 */
import type { CommentaryInputType } from './commentaryInputType';

export interface CommentaryInput {
  message: string;
  type: CommentaryInputType;
  /** @nullable */
  lap?: number | null;
  /** @nullable */
  driverName?: string | null;
}