# lib/api-zod/src/generated/types/raceControlMessage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * F1 Live Telecronaca API
 * OpenAPI spec version: 0.1.0
 */
import type { RaceControlMessageCategory } from './raceControlMessageCategory';
import type { RaceControlMessageFlag } from './raceControlMessageFlag';

export interface RaceControlMessage {
  id: number;
  raceId: number;
  message: string;
  category: RaceControlMessageCategory;
  /** @nullable */
  flag?: RaceControlMessageFlag;
  /** @nullable */
  lap?: number | null;
  timestamp: string;
}